Refactor pr.js to run.js for consistent checking
