+++ .github/workflows/check-links.yml
@@ -43,6 +43,6 @@ jobs:
         run: |
           cd tools/broken-link-checker
           npm ci
-          node pr.js
+          node run.js pr
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN  }}
