+++ tools/broken-link-checker/ignored_paths.json
@@ -1,0 +1,3 @@
+[
+  ".*plugin-development.*"
+]
