+++ tools/broken-link-checker/ignored_targets.json
@@ -1,0 +1,28 @@
+[
+  "http://127.0.0.1:8000/*",
+  "http://127.0.0.1:8001/*",
+  "http://127.0.0.1:8002/*",
+  "http://127.0.0.1:8003/*",
+  "http://127.0.0.1:8004/*",
+  "http://localhost:3000/",
+  "http://localhost:8000/*",
+  "http://localhost:8001/*",
+  "http://localhost:8002/*",
+  "http://localhost:8003/*",
+  "http://localhost:8004/*",
+  "http://localhost:9090/*",
+  "insomnia://*",
+  "insomniad://",
+  "*/hub/*",
+  "https://linkedin.com/*",
+  "https://aws.amazon.com/*",
+  "http://xip.io",
+  "https://kong.127-0-0-1.nip.io/*",
+  "https://kong.127-0-0-1.nip.io",
+  "https://kong.7f000001.nip.io/*",
+  "https://portal.7f000001.nip.io/*",
+  "https://manager.7f000001.nip.io/*",
+  "https://admin.okta.com/*",
+  "https://kong.example.com",
+  "https://kong.127-0-0-1.nip.io/api"
+]
